requirements.txt
paramiko
fastapi
sqlmodel
uvicorn
mysql-connector-python
python-dotenv
alembic